refactor: move trace-map dir path
