+++ server/querier/app/distributed-tracing/common/const.go
@@ -13,13 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package tracemap
 
-import (
-	"github.com/deepflowio/deepflow/server/querier/app/distributed-tracing/model"
-	"github.com/deepflowio/deepflow/server/querier/config"
+package common
+
+const (
+	SUCCESS            = "SUCCESS"
+	FAIL               = "FAIL"
+	INVALID_PARAMETERS = "INVALID_PARAMETERS"
+	INVALID_POST_DATA  = "INVALID_POST_DATA"
+	SERVER_ERROR       = "SERVER_ERROR"
 )
 
-func TraceMap(args model.TraceMap, cfg *config.QuerierConfig) (result model.TraceMapTree, debug interface{}, err error) {
-	return
-}
+const (
+	DATABASE_FLOW_LOG = "flow_log"
+	TABLE_L7_FLOW_LOG = "l7_flow_log"
+	TAG_TRACE_ID      = "trace_id"
+)
+
+const (
+	HEADER_KEY_X_ORG_ID = "X-Org-Id"
+)
