--- server/querier/app/distributed-tracing/service/common/const.go
@@ -1,19 +1,0 @@
-package common
-
-const (
-	SUCCESS            = "SUCCESS"
-	FAIL               = "FAIL"
-	INVALID_PARAMETERS = "INVALID_PARAMETERS"
-	INVALID_POST_DATA  = "INVALID_POST_DATA"
-	SERVER_ERROR       = "SERVER_ERROR"
-)
-
-const (
-	DATABASE_FLOW_LOG = "flow_log"
-	TABLE_L7_FLOW_LOG = "l7_flow_log"
-	TAG_TRACE_ID      = "trace_id"
-)
-
-const (
-	HEADER_KEY_X_ORG_ID = "X-Org-Id"
-)
